Move non_search tests under dedicated non_search tests dir
