+++ tests/non_search/test_auth.py
