+++ tests/non_search/test_geo_stats.py
