+++ tests/non_search/test_pipeline_trigger.py
