+++ tests/unit/app/core/test_pipeline.py
@@ -1,7 +1,7 @@
 from sqlalchemy.orm import Session
 
 from app.core.ingestion.pipeline import generate_pipeline_ingest_input
-from tests.routes.setup_helpers import setup_with_two_docs_one_family
+from tests.non_search.setup_helpers import setup_with_two_docs_one_family
 
 
 def test_generate_pipeline_ingest_input(data_db: Session):
